requirements.txt
playwright==1.42.0
supabase==2.3.7
httpx<0.28.0
openpyxl==3.1.2
requests==2.31.0
python-jwt==4.1.0
tk==0.1.0
python-dotenv
PyQt6
PyQt6-WebEngine
